+++ src/components/Layout.tsx
@@ -3,6 +3,7 @@ import { useStaticQuery, graphql } from 'gatsby';
 import { Helmet } from 'react-helmet';
 import { SubHeader } from './SubHeader';
 import { Header } from './../components/Header';
+import { Footer } from './../components/Footer';
 
 interface LayoutProps {
   children: ReactChild;
@@ -16,6 +17,9 @@ export const Layout = ({ children, showSubHeader }: LayoutProps) => {
         site {
           siteMetadata {
             title
+            description
+            socialImage
+            website
           }
         }
       }
@@ -25,35 +29,21 @@ export const Layout = ({ children, showSubHeader }: LayoutProps) => {
     <>
       <Helmet>
         <meta charSet="utf-8" />
-        <meta property="og:title" content="Grafana Infinity Datasource Plugin" />
-        <meta
-          property="og:description"
-          content="Do infinite things with Grafana. Turn any website into beautiful grafana dashboards. Supports HTML, CSV, JSON, XML & GraphQL documents."
-        />
-        <meta
-          property="og:image"
-          content="https://user-images.githubusercontent.com/153843/92741922-03491380-f377-11ea-9c31-9a744afd3388.png"
-        />
-        <meta property="og:url" content="https://yesoreyeram.github.io/grafana-infinity-datasource" />
+        <meta property="og:title" content={data.site.siteMetadata.title} />
+        <meta property="og:description" content={data.site.siteMetadata.description} />
+        <meta property="og:image" content={data.site.siteMetadata.socialImage} />
+        <meta property="og:url" content={data.site.siteMetadata.website} />
         <meta property="twitter:card" content="summary_large_image" />
         <meta property="twitter:site" content="@yesoreyeram.github.io/grafana-infinity-datasource" />
-        <meta property="twitter:title" content="Grafana Infinity Datasource Plugin." />
-        <meta
-          property="twitter:description"
-          content="Do infinite things with Grafana. Turn any website into beautiful grafana dashboards. Supports HTML, CSV, JSON, XML & GraphQL documents."
-        />
-        <meta
-          property="twitter:image"
-          content="https://user-images.githubusercontent.com/153843/92741922-03491380-f377-11ea-9c31-9a744afd3388.png"
-        />
-        <meta property="" content="" />
-        <meta property="" content="" />
-        <title>Infinity Datasource</title>
-        <link rel="canonical" href="http://mysite.com/example" />
+        <meta property="twitter:title" content={data.site.siteMetadata.title} />
+        <meta property="twitter:description" content={data.site.siteMetadata.description} />
+        <meta property="twitter:image" content={data.site.siteMetadata.socialImage} />
+        <title>{data.site.siteMetadata.title}</title>
       </Helmet>
       <Header title={data.site.siteMetadata.title} />
       {showSubHeader && <SubHeader></SubHeader>}
-      {children}
+      <div>{children}</div>
+      <Footer />
     </>
   );
 };